LCORE-1831: move RedactionResult to core.py
