+++ src/models/config.py
@@ -2423,23 +2423,6 @@ class RedactionConfig(ConfigurationBase):
         return list(self._compiled_patterns)
 
 
-class RedactionResult(BaseModel):
-    """Result of applying PII redaction rules to text.
-
-    Attributes:
-        content: The text after all redaction rules have been applied.
-        redacted: True if at least one rule matched and changed the text.
-        redaction_count: Total number of substitutions made across all rules.
-    """
-
-    model_config = ConfigDict(frozen=True)
-
-    content: str
-    redacted: bool
-    redaction_count: int
-
-
-
 class Configuration(ConfigurationBase):
     """Global service configuration."""
