+++ src/pydantic_ai_lightspeed/capabilities/redaction/__init__.py
@@ -2,13 +2,13 @@
 
 from models.config import (
     RedactionConfig,
-    RedactionResult,
     RedactionRule,
 )
 from pydantic_ai_lightspeed.capabilities.redaction.capability import (
     PiiRedactionCapability,
 )
 from pydantic_ai_lightspeed.capabilities.redaction.core import (
+    RedactionResult,
     redact_text,
 )
